Update: Added new directory with my resume copy!
I have progressed in the research and started somewhere in the path of learning something. Trying something differently and picking up the pace with whatever I am doing. I have created an HTML page of my resume and uploaded it here to edit it in a visually appealing manner later.